Added Kafka, fixed dependencies and dockerfile, added endpoint
