+++ src/tests/api_tests.py
@@ -1,0 +1,8 @@
+import pytest
+import requests
+
+BASE_URL = "http://localhost:8000"  
+
+def dummy_test():
+    pass 
+
