--- tests/api_tests.py
@@ -1,8 +1,0 @@
-import pytest
-import requests
-
-BASE_URL = "http://localhost:8000"  
-
-def dummy_test():
-    pass 
-
